Add files via upload
this project is about sharing file sharing using js file with env and html files.   it will provide a link when you upload the file and you can set the password for it .when you share the link the user might enter the password in order to download or view the file.